% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_brodmanns.R
\name{add_brodmanns}
\alias{add_brodmanns}
\title{Add Brodmann's areas to the data}
\usage{
add_brodmanns(nirsData)
}
\arguments{
\item{data}{The data to which Brodmann's areas will be added. This data should have been imported using the \code{\link{import_nirs}} function.}
}
\value{
The data with Brodmann's areas added. There will be 2 new columns: \code{hemi} and \code{region}. The \code{hemi} column indicates whether the optode is on the left or right hemisphere. The \code{region} column indicates the Brodmann's area.
}
\description{
This function adds Brodmann's areas to the data in a new column. Brodmann's areas can be added to raw nir data or to HbO/HbR data. The assignment of Brodmann's areas is based on the coordinates of the optodes.
}
\examples{
\dontrun{
data <- import_nirs("data.nir")
data <- addBrodmanns(data)
}
}
