% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers.R
\name{add_markers}
\alias{add_markers}
\title{add_markers}
\usage{
add_markers(nirsData)
}
\arguments{
\item{nirsData}{A dataframe containing NIRS data in a long format that has been imported using the import_nirs function.}
}
\description{
This function will add markers to the NIRS data. The data must be in a long format (imported using import_nirs) and markers must be in a separate file with the exact same name as the NIRS file (with .mrk extension).
}
\examples{
# The location of the marker file will be read from the fileName column of the nirsData dataframe.
# The marker file must have the same name as the NIRS file with a .mrk extension.
nirsData <- import_nirs("path/to/nir/file.nir")
nirsData <- add_markers(nirsData)
}
