% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_baseline_correction.R
\name{apply_baseline_correction}
\alias{apply_baseline_correction}
\title{Apply Local Baseline Correction (Robust)}
\usage{
apply_baseline_correction(epoched_data, baseline_duration = 5)
}
\arguments{
\item{epoched_data}{Dataframe from get_event_data().}

\item{baseline_duration}{(numeric) Length of baseline in \emph{SECONDS} (e.g., 5).
This usually matches the 'lead' time you used in get_event_data.}
}
\value{
Dataframe with 'hbo_corr', 'hbr_corr', and 'rel_time'.
}
\description{
Aligns trials to zero using relative time.
}
