% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_mbll.R
\name{apply_mbll}
\alias{apply_mbll}
\title{Apply the Modified Beer-Lambert Law to delta oxy values to calculate the concentration changes of HbO and HbR.}
\usage{
apply_mbll(deltaOds, e_matrix, dpf = 6, l = 2.5)
}
\arguments{
\item{e_matrix}{A matrix with the extinction coefficients of HbO and HbR at the two wavelengths. This should be the output of the \code{get_ext_values} function.}

\item{dpf}{The differential pathlength factor. The default is 6. This can be changed if the dpf is known to be different.}

\item{l}{The distance between the light source and the detector (in cm). The default is 2.5cm. This can be changed if the distance is known to be different.}

\item{delta_od}{A dataframe with the delta_od values. This should be the output of the \code{create_delta_od} function.}
}
\value{
A dataframe with the concentration changes of HbO and HbR.
}
\description{
This function applies the Modified Beer-Lambert Law to delta oxy values to calculate the concentration changes of HbO and HbR. It does so using the extinction coefficients of HbO and HbR at the two wavelengths that were calculated using the \code{get_ext_values} function.
}
\examples{
\dontrun{
# Assuming that the delta_od values have been calculated using the \code{\link{create_delta_od}} function and the extinction coefficients have been calculated using the \code{\link{get_ext_values}} function. dpf is set to 6 and l is set to 2.5.
applyMBLL(delta_ods, e_matrix, dpf = 6, l = 2.5)
}
}
