% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_smar.R
\name{apply_smar}
\alias{apply_smar}
\title{Apply a Sliding Window Motion Artifact Removal algorithm to the fnirs data.}
\usage{
apply_smar(
  nirsData,
  window_size_seconds = 5,
  sample_rate = 2,
  upper_threshold = 0.025,
  lower_threshold = 0.003,
  preserve_all = T
)
}
\arguments{
\item{nirsData}{(dataframe)  NIRS data that has been imported using the \code{\link{import_nirs}} function.}

\item{window_size_seconds}{(numeric) The window size in seconds for the sliding window. The default is 10 (5sec at 2Hz).}

\item{sample_rate}{(numeric) The sample rate of the data. The default is 2.}

\item{upper_threshold}{(numeric) The upper threshold for identifying artifacts. The default is 0.025.}

\item{lower_threshold}{(numeric) The lower threshold for identifying artifacts. The default is 0.003.}

\item{perserve_all}{(logical) If TRUE, each step of the process will be preserved in the data, with _pre_smar (original data) and _pre_smooth (artifacts removed, before overall smoothing) added to the column name. The default is FALSE.}
}
\value{
A dataframe with the sliding window Motion Artifact Removal algorithm applied to the data. Original data is preserved with the suffix "_pre_smar" added to the column names. The nirValue column is replaced with the smoothed values.
This function is not ready for production use. It is a work in progress.
}
\description{
This function will apply a sliding window Motion Artifact Removal algorithm to the fnirs data. The window size, sample rate, and thresholds can be specified in the function.

There are two steps to the algorithm:
\enumerate{
\item Identify artifacts based on thresholds. The differences between consecutive values are calculated, and if the difference is greater than the upper threshold or less than the lower threshold, the data point is identified as an artifact. It is then replaced with an interpolated value.
\item Apply simple smoothing to the data using a manual sliding window - centered average. The centered average is calculated for each data point using the window size specified.
}

The output is a dataframe with the sliding window Motion Artifact Removal algorithm applied to the data, and the original data is preserved with the suffix "_pre_smar" added to the column name.
}
