% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_delta_ods.R
\name{create_delta_ods}
\alias{create_delta_ods}
\title{Calculate change in optical density (delta od) from NIRS data}
\usage{
create_delta_ods(nirsData, reference = "baseline")
}
\arguments{
\item{nirsData}{(DATAFRAME) This should be in the format of the output from the function \code{\link{import_nirs}}.}

\item{reference}{(STRING) specifying the reference value to use for calculating delta od. This can be either 'baseline' or 'mean'. If 'baseline', the baseline value for each channel is used as the reference value. If 'mean', the mean value for each channel is used as the reference value. Default is 'baseline'.}
}
\value{
a data frame which adds the delta od values for each channel to a new column.
}
\description{
Calculates the change in oxygen density (delta od) for each channel from NIRS data, based on a reference value.
}
\examples{
\dontrun{
nirsData <- import_nirs("path/to/nirs/data.nir")
# Create the delta ODs
deltaOds <- create_delta_ods(nirsData, reference = "baseline")
}
}
\seealso{
\code{\link{import_nirs}}, \code{\link{apply_mbll}}
}
