% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_utils.R
\name{create_regressor}
\alias{create_regressor}
\title{Create Design Matrix Regressor}
\usage{
create_regressor(times, event_times, hrf)
}
\arguments{
\item{times}{Vector of time points from your data.}

\item{event_times}{Vector of specific event onset times (in seconds).}

\item{hrf}{The HRF vector from get_canonical_hrf().}
}
\value{
A numeric vector of the same length as 'times'.
}
\description{
Converts event markers into a continuous time-series regressor by convolving
a stick function with the canonical HRF. This handles the \emph{overlap} between trials.
}
