% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_utils.R
\name{get_canonical_hrf}
\alias{get_canonical_hrf}
\title{Generate Canonical HRF (Double Gamma)}
\usage{
get_canonical_hrf(tr, duration = 32)
}
\arguments{
\item{tr}{Time repetition (sampling interval in seconds). E.g., 0.5 for 2Hz.}

\item{duration}{Length of the HRF kernel in seconds (default 32s).}
}
\value{
A numeric vector representing the HRF.
}
\description{
Generates a standard double-gamma Hemodynamic Response Function.
This "ideal" wave shape is used to model what the brain response should look like.
}
