% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ext_values.R
\name{get_ext_values}
\alias{get_ext_values}
\title{Get the extinction coefficients of HbO and HbR at the two wavelengths.}
\usage{
get_ext_values(lambda1 = 730, lambda2 = 850, table = "wray")
}
\arguments{
\item{lambda1}{The first wavelength.}

\item{lambda2}{The second wavelength.}

\item{table}{Which table of extinction coefficients to use. Options are "wray", "takatani", "moaveni", "gratzer" and "cope".}
}
\value{
a matrix with the extinction coefficients of HbO and HbR at the two wavelengths, which can be used with the \code{applyMBLL} function.
}
\description{
This function returns the extinction coefficients of HbO and HbR at the two wavelengths. The extinction coefficients are taken from the literature.
}
