% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_nirs.R
\name{import_nirs}
\alias{import_nirs}
\title{Import NIRS data}
\usage{
import_nirs(nirFile)
}
\arguments{
\item{nirFile}{Path to the .nir file}
}
\value{
A dataframe with the data in a long format with the following columns: t, startTime, fileName, dataRow, optode, freq, nirValue.
\itemize{
\item t: time in seconds from the start of the recording
\item startTime: the start time of the recording
\item fileName: the name of the file (full path)
\item dataRow: the row number of the data as it is in the original .nir file. The first row of the data is the first row after the baseline data. You will see that this matches the values of the first column in COBI Studio .nir file.
\item optode: the optode number (e.g., 1-16)
\item freq: the frequency (e.g., 730, 850)
\item nirValue: the value of the NIR data for that optode and frequency at that time point.
}
}
\description{
This function will import a .nir file, identify the baseline values and return a dataframe with the data in a long format.
}
\details{
This function will import a .nir file, identify the baseline values and return a dataframe with the data in a long format.
}
\examples{
nirsData <- import_nirs("path/to/nir/file.nir")
}
