% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ambient.R
\name{remove_ambient}
\alias{remove_ambient}
\title{Remove ambient light channel from NIRS data}
\usage{
remove_ambient(nirsData)
}
\arguments{
\item{nirsData}{A dataframe containing NIRS data in a long format that has been imported using the import_nirs function.}
}
\description{
This function will remove the ambient light channel from the NIRS data. It is not required to do this, but it can make the data easier to work with.
}
\examples{
nirsData <- import_nirs("path/to/nir/file.nir")
nirsData <- remove_ambient(nirsData)
}
