% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_channels.R
\name{remove_channels}
\alias{remove_channels}
\title{Remove channels from NIRS data}
\usage{
remove_channels(nirsData, channels)
}
\arguments{
\item{nirsData}{A dataframe containing NIRS data in a long format that has been imported using the import_nirs function.}

\item{channels}{A vector of channel names to remove from the NIRS data. The channel names must match the names in the \emph{optode} column of the nirs data (i.e., 1-16).}
}
\value{
A dataframe with the specified channels removed.
}
\description{
This function will remove the specified channels from the NIRS data. This can be useful if you have channels that are not required for your analysis or if channels are saturated or weak, or noisy.
}
\examples{
nirsData <- import_nirs("path/to/nir/file.nir")
problem_channels <- c("1", "2", "3")
nirsData <- remove_channels(nirsData, problem_channels)
}
\seealso{
\code{\link{import_nirs}}
}
