% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_plot.R
\name{signal_plot}
\alias{signal_plot}
\title{Generate a plot to view the raw signal data from each optode.}
\usage{
signal_plot(nirsData, path = NULL)
}
\arguments{
\item{nirsData}{(dataframe)  NIRS data that has been imported using the \code{\link{import_nirs}} function.}
}
\value{
A plot with the raw signal data from each optode.
}
\description{
This function will generate a plot to view the raw signal data from each optode. The output is a plot with the signal data on the y-axis and the time on the x-axis. The plot is meant as a quick way to check the quality of the signal data. Can be used in conjunction with the \code{\link{signal_summary}} function to determine low quality channels.
}
\examples{
\dontrun{
# To generate a signal plot from a single NIRS file
nirsData <- import_nirs("path/to/nir/file.nir")
signalPlot <- signal_plot(nirsData)

# To generate a signal plot from multiple NIRS files which have been imported using the import_nirs function
nirsData <- lapply(nirsFiles, import_nirs, path = myPath)
signalPlot <- lapply(nirsData, signal_plot)
}

}
\seealso{
\code{\link{import_nirs}}
}
