% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_lowpass.R
\name{apply_lowpass}
\alias{apply_lowpass}
\title{Apply lowpass filter to fnirs data.}
\usage{
apply_lowpass(nirData, freq = 0.1, w = 20)
}
\arguments{
\item{freq}{(numeric) The frequency of the lowpass filter. The default is 0.1 (10Hz).}

\item{w}{(numeric) The window of the lowpass filter. The default is 20.}

\item{oxyData}{(dataframe)  NIRS data that has been imported using the \code{\link{import_nirs}} function.}
}
\value{
A dataframe with the lowpass filter applied to the data. Original data is preserved with the suffix "_pre_lp" added to the column names. The nirValue column is replaced with the lowpass filtered values.
}
\description{
This function will apply a lowpass filter to the fnirs data. The frequency of the filter and the window can be specified in the function. The output is a dataframe with the lowpass filter applied to the data, and the original data is preserved with the suffix "_pre_lp" added to the column names. The function uses a fir1 filter from the gsignal package, with a filter length of 2. The filter is scaled by dividing by the sum of the coefficients, and the starting point values are created using the filter_zi function. The filter is applied to the data, and the delay is calculated using the grpdelay function. The delay is then compensated for by shifting the data. The output is a dataframe with the lowpass filter applied to the data.
}
\examples{
\dontrun{
# Assuming that the fnirs data has been imported using the \code{\link{import_nirs}} function.
apply_lowpass(nirData, freq = 0.1, w = 20)
}
}
