% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_mbll.R
\name{apply_mbll}
\alias{apply_mbll}
\title{Apply the Modified Beer Lambert Law to the change in optical density data}
\usage{
apply_mbll(deltaOdData, extMatrix, pathLength = 2.5, dpf = 6)
}
\arguments{
\item{deltaOdData}{a data frame containing the delta od values for each channel. This should be in the format of the output from the function \code{\link{create_delta_ods}}.}

\item{extMatrix}{a numeric matrix containing the extinction coefficients for each wavelength and chromophore. This should be in the format of the output from the function \code{\link{create_ematrix}}.}

\item{pathLength}{a numeric value specifying the pathlength of the light through the tissue. This should be in cm. The default value is 2.5 cm.}

\item{dpf}{a numeric value specifying the differential pathlength factor. The default value is 6.}
}
\value{
a data frame which adds the HbO and HbR values for each channel to new columns.
}
\description{
Applies the Modified Beer Lambert Law to the change in optical density data to calculate the change in concentration of oxygenated and deoxygenated hemoglobin.
}
\examples{
data <- read_nirs_data("data.csv")
data <- create_delta_ods(data, reference = "baseline")
extMatrix <- create_ematrix(lambda1 = 730, lambda2 = 850, table= "wray")
data <- apply_mbll(data, ematrix)
}
