% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_data.R
\name{get_event_data}
\alias{get_event_data}
\title{Get event data using markers}
\usage{
get_event_data(data, eventStart, eventEnd = NULL, eventName, lead = 0, lag = 0)
}
\arguments{
\item{data}{A data frame with the fnirs data. It should have been imported using the \code{\link{import_nirs}} function and had markers added using the \code{\link{add_markers}} function.}

\item{eventStart}{The name of the start marker for discrete events/trials in the data. This should be a string and it should match a value in the marker column.}

\item{eventEnd}{The name of the end marker for discrete events/trials in the data. This should be a string and it should match a value in the marker column. If you do not set an end marker, the function will assume that the end marker is the same as the start marker. This is useful if you have a single marker for each event. If you do not set an end marker, you will need to set the lead and lag parameters. If you do not set the lead and lag parameters for a single marker event, the function will default to a lead of 6 samples and a lag of 10 samples. To calculate seconds, divide by the sample rate (e.g. at 2 Hz would be a value of 10 for 5 seconds).}

\item{eventName}{The name of the block. This is for the convenience of the researcher, to be able to distinguish between different blocks of trials, or different tasks. This should be a string and will be added as a column to the data.}

\item{lead}{The number of data points to lead the start marker by. This is useful because of the lead/lag of fnirs responses to events. The default is 0.  To calculate seconds, divide by the sample rate (e.g. at 2 Hz would be a value of 10 for 5 seconds).}

\item{lag}{The number of data points to lag the end marker by. This is useful because of the lead/lag of fnirs responses to events. The default is 0. To calculate seconds, divide by the sample rate (e.g. at 2 Hz would be a value of 10 for 5 seconds).}
}
\value{
A list of data frames, one for each event. Each data frame will have a trial column that specifies the event number and a block column that specifies the block name.
}
\description{
This function gets the fnirs data for events in the study by identifying start and end markers in the data. It is useful if you have set up the study with automated (serial port) markers that specify when events happen in the data. Before running this function, you should also run the \code{\link{add_markers}} function to bind the markers to the data.

The function will look for matching start and end markers. It will then extract the data between the start and end markers and return a list of data frames, one for each event. You can specify a lead or lag value to extract data before or after the marker values. The data frames will have a trial column that specifies the event number. The function will also add an event_name to the data that specifies the event name (inputted as a parameter). If an end marker is not specified, the function will assume that the end marker is the same as the start marker. This is useful if you have a single marker for each event.
}
